SP-ization (#1562)
* Start on pcg builder

* Add tests and some implementation for pcg builder

* Add pcg tests, make dtgen constructors explicit to fix bug

* Add remainder of PCG tests

* Fix build issues in local-execution

* Format

* Address Reyna comments, add topological_order function for PCG

* Pre multidigraph refactor

* Removing visitable from sp code

* Add open dataflow graph, start to replace pcg dataflow graph

* Start refactoring substitutions

* Add utility functions to support pattern matching

* Spization Algorithm

* Small fix

* Pre-refactor inputs

* Fix proj url

* implementation + tests for get_longest_path_lengths

* minor fix

* Simple sp-ization refactor

* Naive Sp-ization algo refactor

* Implementation and Tests for dependency-preserving spization

* Formatting

* Added testcase for NASNET-A like architecture

* Minor Formatting

* Get back to substitutions, now with unordered graph inputs

* Spization and adjacent functions

* Spization and adjacent changes

* Additional function + bug testing

* Fixes after PR review

* Get substitutions building

* substitutions-tests now builds

* Fix bug in filter, pass some initial substitution tests

* Add tests for fmt::to_string, fix some substitutions bugs

* Pass initial unit tests for find_pattern_matches

* Start on unit tests for pcg pattern

* Updates

* PR Fixes + additional functionality

* Minor refactoring

* Metrics and Testing Tools for SP-ization

* Minor changes

* initial implementation for cost_aware_barrier_sync

* Pass initial test for find_pattern_matches

* Fix small build issue in tests

* Format

* Sync tests in CI with tests in proj

* Fix minor build errors in kernels and local-execution

* Format

* Remove outdated code

* More outdated code removal

* More cleanup, add test for sp decomposition

* Pull apart containers.h

* More sp testing and fixes

* Break up graph algorithms.h

* Pre- full SP algo commit

* Add initial implementation and tests for cbc decomposition and inverse line graph

* Pass test for get_inverse_line_graph

* Add new multidigraph

* Fix get_inverse_line_graph to return a MultiDiGraph instead of a DiGraph

* Fix

* Add tests for parallel and series reduction finding

* Add really rough implementation of valdez sp decomposition

* Fixes + sp-ization benchmarking

* Fix local-execution build

* Add implementations and tests for applying series/parallel reductions

* Additional SP-BenchMarking

* Formatting

* Refactoring

* Format

* Clean up sp decomposition interface and tests

* Format

* Add comments for top-level substitutions functions, add proj doxygen support

* Fix proj invocation in CI

* Sp-ization reorganization

* Refactoring + bringing up-to-date with DataFlow-graph interface

* minor fix

* Additional is_acyclic testing

* Fixes

* Additional SP-ization metric + Transitive Reduction Fix + DiGraph from SerialParallelDecomposition

* Benchmarking updates

* formatting

* fixes

* spization updates

* fixes

* removed delete_node

* added docstrings

* minor

* v1 of flexible algo

* fmt

* fixed bug

* updates

* minor

* addressed comments

* minor

* Tiny formatting fix

---------

+++ lib/utils/test/src/utils/graph/series_parallel/binary_sp_decomposition_tree/binary_sp_decomposition_tree.cc
@@ -15,7 +15,9 @@ TEST_SUITE(FF_TEST_SUITE) {
       return BinarySPDecompositionTree{BinaryParallelSplit{lhs, rhs}};
     };
 
-    auto make_leaf = [](int n) { return BinarySPDecompositionTree{Node{n}}; };
+    auto make_leaf = [](size_t n) {
+      return BinarySPDecompositionTree{Node{n}};
+    };
 
     SUBCASE("leaf only") {
       BinarySPDecompositionTree input = make_leaf(5);
